# OPS/openapi/300-ops-api.yaml
openapi: 3.1.0
info:
  title: 300 OPS Platform API
  description: |
    300 OPS is an engineering platform for smart buildings. This API provides
    comprehensive access to project management, engineering evaluation,
    prescription management, procurement, document generation, integrations,
    and AI services.
  version: 1.0.0
  contact:
    name: 300 OPS Support
    email: api@300ops.com
  license:
    name: Proprietary
    url: https://300ops.com/license
servers:
  - url: https://api.300ops.com/v1
    description: Production server
  - url: https://staging-api.300ops.com/v1
    description: Staging server
security:
  - BearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate user
      description: Authenticate a user and return JWT tokens.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: Refresh an expired access token using a refresh token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /equipment:
    get:
      tags: [Equipment]
      summary: List equipment instances
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/FilterParam'
        - name: room_id
          in: query
          schema:
            type: string
            format: uuid
        - name: subsystem_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of equipment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedEquipment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Equipment]
      summary: Create equipment instance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentCreate'
      responses:
        '201':
          description: Equipment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeEquipment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
  /equipment/{id}:
    parameters:
      - $ref: '#/components/parameters/UUIDPathParam'
    get:
      tags: [Equipment]
      summary: Get equipment by ID
      responses:
        '200':
          description: Equipment found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeEquipment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Equipment]
      summary: Update equipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentUpdate'
      responses:
        '200':
          description: Equipment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeEquipment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Equipment]
      summary: Delete equipment
      responses:
        '204':
          description: Equipment deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ─────────────────────────────────────────────────────────
  # EQUIPMENT LIBRARY
  # ─────────────────────────────────────────────────────────
  /equipment-library:
    get:
      tags: [Equipment Library]
      summary: Browse equipment library
      description: Search and filter the master equipment catalog.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/FilterParam'
        - name: category
          in: query
          schema:
            type: string
        - name: manufacturer
          in: query
          schema:
            type: string
        - name: search
          in: query
          description: Full-text search across name, description, and tags
          schema:
            type: string
      responses:
        '200':
          description: Equipment library entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedEquipmentLibrary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Equipment Library]
      summary: Add equipment to library
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentLibraryCreate'
      responses:
        '201':
          description: Equipment added to library
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeEquipmentLibrary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
  /equipment-library/{id}:
    parameters:
      - $ref: '#/components/parameters/UUIDPathParam'
    get:
      tags: [Equipment Library]
      summary: Get library equipment by ID
      responses:
        '200':
          description: Library equipment found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeEquipmentLibrary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Equipment Library]
      summary: Update library equipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentLibraryUpdate'
      responses:
        '200':
          description: Library equipment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeEquipmentLibrary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Equipment Library]
      summary: Remove from library
      responses:
        '204':
          description: Library equipment removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /engineering-rules:
    get:
      tags: [Engineering Rules]
      summary: List engineering rules
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/FilterParam'
        - name: rule_type
          in: query
          schema:
            type: string
        - name: is_active
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of engineering rules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedEngineeringRules'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Engineering Rules]
      summary: Create engineering rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EngineeringRuleCreate'
      responses:
        '201':
          description: Engineering rule created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeEngineeringRule'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
  /engineering-rules/{id}:
    parameters:
      - $ref: '#/components/parameters/UUIDPathParam'
    get:
      tags: [Engineering Rules]
      summary: Get engineering rule by ID
      responses:
        '200':
          description: Engineering rule found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeEngineeringRule'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Engineering Rules]
      summary: Update engineering rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EngineeringRuleUpdate'
      responses:
        '200':
          description: Engineering rule updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeEngineeringRule'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Engineering Rules]
      summary: Delete engineering rule
      responses:
        '204':
          description: Engineering rule deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ─────────────────────────────────────────────────────────
  # PROPOSALS
  # ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login
  parameters:
    UUIDPathParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique resource identifier
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
    SortParam:
      name: sort
      in: query
      schema:
        type: string
        example: created_at:desc
      description: Sort field and direction (field:asc or field:desc)
    FilterParam:
      name: filter
      in: query
      schema:
        type: string
      description: JSON-encoded filter object
  schemas:
...
    Meta:
      type: object
      required: [version, request_id, timestamp]
      properties:
        version:
          type: string
          example: "1.0"
        request_id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        timestamp:
          type: string
          format: date-time
        confidence:
          type: number
          minimum: 0
          maximum: 1
          example: 0.94
        needs_validation:
          type: boolean
          default: false
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            total_pages:
              type: integer
            has_next:
              type: boolean
            has_prev:
              type: boolean
    Links:
      type: object
      required: [self]
      properties:
        self:
          type: string
          format: uri
        related:
          type: object
          additionalProperties:
            type: string
            format: uri
    Envelope:
      type: object
      required: [data, meta, links]
      properties:
        data:
          type: object
        meta:
          $ref: '#/components/schemas/Meta'
        links:
          $ref: '#/components/schemas/Links'

    # ───────────────────────────────────────────────
    # ERROR SCHEMAS
    # ───────────────────────────────────────────────
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        field:
          type: string
        details:
          type: object
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'

    # ───────────────────────────────────────────────
    # AUTH SCHEMAS
    # ───────────────────────────────────────────────
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 6
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string

    # ───────────────────────────────────────────────
    # PROFILE SCHEMAS
    # ───────────────────────────────────────────────
...
    Equipment:
      type: object
      required: [id, name, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        room_id:
          type: string
          format: uuid
        subsystem_id:
          type: string
          format: uuid
        library_reference_id:
          type: string
          format: uuid
        quantity:
          type: integer
          default: 1
        status:
          type: string
          enum: [specified, approved, ordered, delivered, installed, commissioned]
        specifications:
          type: object
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    EquipmentCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        room_id:
          type: string
          format: uuid
        subsystem_id:
          type: string
          format: uuid
        library_reference_id:
          type: string
          format: uuid
        quantity:
          type: integer
          default: 1
        status:
          type: string
          enum: [specified, approved, ordered, delivered, installed, commissioned]
        specifications:
          type: object
        metadata:
          type: object
    EquipmentUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        room_id:
          type: string
          format: uuid
        subsystem_id:
          type: string
          format: uuid
        library_reference_id:
          type: string
          format: uuid
        quantity:
          type: integer
        status:
          type: string
          enum: [specified, approved, ordered, delivered, installed, commissioned]
        specifications:
          type: object
        metadata:
          type: object
    EnvelopeEquipment:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Equipment'
    PaginatedEquipment:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Equipment'

    # ───────────────────────────────────────────────
    # EQUIPMENT LIBRARY SCHEMAS
    # ───────────────────────────────────────────────
    EquipmentLibrary:
      type: object
      required: [id, name, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        category:
          type: string
        manufacturer:
          type: string
        model_number:
          type: string
        datasheet_url:
          type: string
          format: uri
        specifications:
          type: object
        certifications:
          type: array
          items:
            type: string
        estimated_price:
          type: number
        currency:
          type: string
          default: EUR
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    EquipmentLibraryCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        manufacturer:
          type: string
        model_number:
          type: string
        datasheet_url:
          type: string
          format: uri
        specifications:
          type: object
        certifications:
          type: array
          items:
            type: string
        estimated_price:
          type: number
        currency:
          type: string
          default: EUR
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
    EquipmentLibraryUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        manufacturer:
          type: string
        model_number:
          type: string
        datasheet_url:
          type: string
          format: uri
        specifications:
          type: object
        certifications:
          type: array
          items:
            type: string
        estimated_price:
          type: number
        currency:
          type: string
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
    EnvelopeEquipmentLibrary:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/EquipmentLibrary'
    PaginatedEquipmentLibrary:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/EquipmentLibrary'

    # ───────────────────────────────────────────────
    # ENGINE SCHEMAS
    # ───────────────────────────────────────────────
...
    EngineeringRule:
      type: object
      required: [id, name, rule_type, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        rule_type:
          type: string
        condition:
          type: object
        action:
          type: object
        priority:
          type: integer
        is_active:
          type: boolean
          default: true
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    EngineeringRuleCreate:
      type: object
      required: [name, rule_type]
      properties:
        name:
          type: string
        description:
          type: string
        rule_type:
          type: string
        condition:
          type: object
        action:
          type: object
        priority:
          type: integer
        is_active:
          type: boolean
          default: true
        metadata:
          type: object
    EngineeringRuleUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        rule_type:
          type: string
        condition:
          type: object
        action:
          type: object
        priority:
          type: integer
        is_active:
          type: boolean
        metadata:
          type: object
    EnvelopeEngineeringRule:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/EngineeringRule'
    PaginatedEngineeringRules:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/EngineeringRule'
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unprocessable:
      description: Unprocessable Entity - validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'